embed fs working
