+++ cmd/main.go
@@ -16,6 +16,7 @@ import (
 	"github.com/teal-fm/piper/db"
 	"github.com/teal-fm/piper/oauth"
 	"github.com/teal-fm/piper/oauth/atproto"
+	pages "github.com/teal-fm/piper/pages"
 	apikeyService "github.com/teal-fm/piper/service/apikey"
 	"github.com/teal-fm/piper/service/musicbrainz"
 	"github.com/teal-fm/piper/service/spotify"
@@ -31,7 +32,7 @@ type application struct {
 	mbService         *musicbrainz.MusicBrainzService
 	atprotoService    *atproto.ATprotoAuthService
 	playingNowService *playingnow.PlayingNowService
-	pages             *Pages
+	pages             *pages.Pages
 }
 
 // JSON API handlers
@@ -106,7 +107,7 @@ func main() {
 		spotifyService:    spotifyService,
 		atprotoService:    atprotoService,
 		playingNowService: playingNowService,
-		pages:             NewPages(true),
+		pages:             pages.NewPages(false),
 	}
 
 	trackerInterval := time.Duration(viper.GetInt("tracker.interval")) * time.Second
